+++ cpp_version/pmugg/network/net_transition.cpp
@@ -28,10 +28,10 @@ NetTransition* NetTransition::import(const Json& json, Shape3D* shape) {
         endNetworks.push_back(endNetwork);
     }
 
-    int vertexSize = startNetworks[0]->getBVertices().size();
-    int halfEdgeSize = startNetworks[0]->getBHalfEdges().size();
-    int faceSize = startNetworks[0]->getBFaces().size();
-    for (int i = 1; i < startNetworks.size(); i++) {
+    size_t vertexSize = startNetworks[0]->getBVertices().size();
+    size_t halfEdgeSize = startNetworks[0]->getBHalfEdges().size();
+    size_t faceSize = startNetworks[0]->getBFaces().size();
+    for (size_t i = 1; i < startNetworks.size(); i++) {
         if (startNetworks[i]->getBVertices().size() != vertexSize) {
             throw std::runtime_error("Boundary vertex mismatch");
         }
